Const Removal
Removed the const for all the elements being stored by the ds_array
module. The consts were originally in to ensure that the caller knew
that elements being stored in the ds_array module will not be changed
by the ds_array module, however this is not true as the caller will
sooner or later request the item back from the array and should be able
to modify the item.

Storing immutable items is not very useful.